ai_app_backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.41
pydantic==2.11.7
langchain-openai==0.3.28
langchain-community==0.3.27
langchain-core==0.3.72
python-dotenv==1.1.1